utils: refactor format_parse_tree module
